+++ src/main/java/com/github/nthykier/debpkg/deb822/dialects/Deb822DialectAptSourcesParserDefinition.java
@@ -1,0 +1,10 @@
+package com.github.nthykier.debpkg.deb822.dialects;
+
+import com.github.nthykier.debpkg.deb822.AbstractDeb822ParserDefinition;
+
+public class Deb822DialectAptSourcesParserDefinition extends AbstractDeb822ParserDefinition {
+
+    public Deb822DialectAptSourcesParserDefinition() {
+        super(Deb822DialectAptSourcesFileType.INSTANCE);
+    }
+}